+++ Source/Lib/Encoder/ASM_SSE4_1/Quant_sse4_1.c
@@ -9,7 +9,7 @@
 #include <immintrin.h>
 
 static void quant_deadzone_sse4_1(uint16_t* buf, uint32_t size, int8_t gtli) {
-    uint16_t clamp = ((uint16_t)(~(uint16_t)0) << gtli) & (~BITSTREAM_MASK_SIGN);
+    uint16_t clamp = (uint16_t)(((unsigned)(uint16_t)(~(uint16_t)0) << gtli) & ~(unsigned)BITSTREAM_MASK_SIGN);
     uint32_t coeff_idx = 0;
 
     __m128i zero = _mm_setzero_si128();
